requirements.txt
Flask
flask-cors
Pillow
tensorflow
tensorflow-hub
kagglehub